Logic

+++ IAA_HA/src/de/nordakademie/action/core/AuthorAction.java
@@ -31,6 +31,10 @@ public class AuthorAction extends ActionSupport implements Action
 
 
 	public String add(){
+		if(authorManager.isAlreadyAvailable(addAuthor)){
+			addFieldError("addAuthor", "Autor existiert bereits");
+			return INPUT;
+		}
 		authorManager.save(addAuthor);
 		return "addAuthor";
 	}
@@ -42,6 +46,10 @@ public class AuthorAction extends ActionSupport implements Action
 	}
 	
 	public String save() {
+		if(authorManager.isAuthorReferenced(resultAuthor)){
+			addFieldError("editAuthor", "Autor wird refernziert");
+			return INPUT;
+		}
 		authorManager.save(editAuthor);
 		authorManager.delete(resultAuthor.getName());
 		return "save";
